Move PackageJson type to own file
