+++ src/utils/AppTypes.ts
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2022 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
+ */
+
+interface ObjectContainingOptionalStrings {
+    [index: string]: string | undefined;
+}
+
+export interface PackageJson {
+    name: string;
+    version: string;
+
+    // Several optional properties
+    author?: string;
+    bin?: ObjectContainingOptionalStrings | string;
+    dependencies?: ObjectContainingOptionalStrings;
+    description?: string;
+    homepage?: string;
+    devDependencies?: ObjectContainingOptionalStrings;
+    displayName?: string;
+    engines?: ObjectContainingOptionalStrings;
+    files?: readonly string[];
+    license?: string;
+    main?: string;
+    peerDependencies?: ObjectContainingOptionalStrings;
+    repository?: {
+        type: string;
+        url: string;
+    };
+    scripts?: ObjectContainingOptionalStrings;
+}
